@@ -3,6 +3,6 @@ import pytest
 pytest_plugins = "pytester"
 
 
-@pytest.fixture()
+@pytest.fixture
 def message_template():
     return "Fixture name: {}, location: {}.py"
